+++ com.unity.ml-agents/Tests/Editor/MLAgentsEditModeTest.cs
@@ -5,6 +5,7 @@ using System.Reflection;
 using System.Collections.Generic;
 using MLAgents.Sensors;
 using MLAgents.Policies;
+using MLAgents.SideChannels;
 
 namespace MLAgents.Tests
 {
@@ -194,7 +195,7 @@ namespace MLAgents.Tests
             Assert.AreEqual(0, aca.EpisodeCount);
             Assert.AreEqual(0, aca.StepCount);
             Assert.AreEqual(0, aca.TotalStepCount);
-            Assert.AreNotEqual(null, aca.FloatProperties);
+            Assert.AreNotEqual(null, aca.GetSideChannel<FloatPropertiesChannel>());
 
             // Check that Dispose is idempotent
             aca.Dispose();
@@ -205,10 +206,10 @@ namespace MLAgents.Tests
         [Test]
         public void TestAcademyDispose()
         {
-            var floatProperties1 = Academy.Instance.FloatProperties;
+            var floatProperties1 = Academy.Instance.GetSideChannel<FloatPropertiesChannel>();
             Academy.Instance.Dispose();
 
-            var floatProperties2 = Academy.Instance.FloatProperties;
+            var floatProperties2 = Academy.Instance.GetSideChannel<FloatPropertiesChannel>();
             Academy.Instance.Dispose();
 
             Assert.AreNotEqual(floatProperties1, floatProperties2);